Remove dependency on keycloak object in user store module Update vuex module unit test structure
