+++ app/frontend/tests/unit/store/modules/auth.getters.spec.js
