+++ app/frontend/tests/unit/store/modules/user.actions.spec.js
@@ -1,71 +1,76 @@
 import { cloneDeep } from 'lodash';
 import { createLocalVue } from '@vue/test-utils';
-import Vue from 'vue';
 import Vuex from 'vuex';
 
 import userService from '@/services/userService';
 import userStore from '@/store/modules/user';
 
+const localVue = createLocalVue();
+localVue.use(Vuex);
+
 const zeroUuid = '00000000-0000-0000-0000-000000000000';
 
-describe('user.js - getUserAcronyms action', () => {
-  const spy = jest.spyOn(userService, 'getUserAcronyms');
-  let store;
+describe('user actions', () => {
+  describe('getUserAcronyms', () => {
+    const spy = jest.spyOn(userService, 'getUserAcronyms');
+    let store;
 
-  beforeEach(() => {
-    const localVue = createLocalVue();
-    localVue.use(Vuex);
-
-    Object.defineProperty(Vue.prototype, '$keycloak', {
-      configurable: true, // Needed to allow deletions later
-      get() {
-        return {
-          authenticated: true,
-          ready: true,
-          subject: zeroUuid
-        };
-      }
+    beforeEach(() => {
+      spy.mockClear();
+      store = new Vuex.Store(cloneDeep(userStore));
     });
 
-    store = new Vuex.Store(cloneDeep(userStore));
-    spy.mockClear();
-  });
+    it('returns acronym details from the api service call', async () => {
+      const mockAcronyms = [
+        { acronym: 'test', owner: false },
+        { acronym: 'other', owner: false }
+      ];
+      userService.getUserAcronyms.mockResolvedValue({ data: mockAcronyms });
+      store.registerModule('auth', {
+        namespaced: true,
+        getters: {
+          authenticated: () => true,
+          subject: () => zeroUuid
+        }
+      });
 
-  afterEach(() => {
-    if (Vue.prototype.$keycloak) {
-      delete Vue.prototype.$keycloak;
-    }
-  });
+      await store.dispatch('getUserAcronyms');
 
-  it('returns acronym details from the api service call', async () => {
-    const mockAcronyms = [
-      { acronym: 'test', owner: false },
-      { acronym: 'other', owner: false }
-    ];
-    userService.getUserAcronyms.mockResolvedValue({ data: mockAcronyms });
+      expect(store.getters.acronyms).toEqual(mockAcronyms);
+      expect(spy).toHaveBeenCalledTimes(1);
+    });
 
-    await store.dispatch('getUserAcronyms');
+    it('handles an empty array', async () => {
+      userService.getUserAcronyms.mockResolvedValue({ data: [] });
+      store.registerModule('auth', {
+        namespaced: true,
+        getters: {
+          authenticated: () => true,
+          subject: () => zeroUuid
+        }
+      });
 
-    expect(store.getters.acronyms).toEqual(mockAcronyms);
-    expect(spy).toHaveBeenCalledTimes(1);
-  });
+      await store.dispatch('getUserAcronyms');
 
-  it('handles an empty array', async () => {
-    userService.getUserAcronyms.mockResolvedValue({ data: [] });
+      expect(store.getters.acronyms).toEqual([]);
+      expect(spy).toHaveBeenCalledTimes(1);
+    });
 
-    await store.dispatch('getUserAcronyms');
+    it('handles a blank return', async () => {
+      const errMsg = 'test error message';
+      userService.getUserAcronyms.mockRejectedValue(errMsg);
+      store.registerModule('auth', {
+        namespaced: true,
+        getters: {
+          authenticated: () => true,
+          subject: () => zeroUuid
+        }
+      });
 
-    expect(store.getters.acronyms).toEqual([]);
-    expect(spy).toHaveBeenCalledTimes(1);
-  });
+      await store.dispatch('getUserAcronyms');
 
-  it('handles a blank return', async () => {
-    const errMsg = 'test error message';
-    userService.getUserAcronyms.mockRejectedValue(errMsg);
-
-    await store.dispatch('getUserAcronyms');
-
-    expect(store.getters.acronyms).toEqual([]);
-    expect(spy).toHaveBeenCalledTimes(1);
+      expect(store.getters.acronyms).toEqual([]);
+      expect(spy).toHaveBeenCalledTimes(1);
+    });
   });
 });
